some utils moved
